fix(compiler): extract sub program compilation
