+++ packages/compiler/src/index.ts
@@ -1,19 +1,5 @@
-import type {
-	CompiledModule,
-	CompileInput,
-	CompileOptions,
-	CompileResult,
-	CompilerCache,
-	FunctionMetadata,
-	FunctionMetadataLookup,
-	FunctionTypeRegistry,
-	ValidatedAST,
-	ValidatedConstantsAST,
-	ValidatedFunctionAST,
-	ValidatedModuleAST,
-	ValidatedPrototypeAST,
-} from '@8f4e/compiler-spec';
-import { DEFAULT_HOST_IMPORT_MODULE_NAME, ErrorCode, GLOBAL_ALIGNMENT_BOUNDARY } from '@8f4e/compiler-spec';
+import type { CompileInput, CompileOptions, CompileResult } from '@8f4e/compiler-spec';
+import { DEFAULT_HOST_IMPORT_MODULE_NAME } from '@8f4e/compiler-spec';
 import {
 	call,
 	createCodeSection,
@@ -37,266 +23,41 @@ import {
 	WASM_MEMORY_PAGE_SIZE,
 	WASM_TYPE_I32,
 } from '@8f4e/compiler-wasm-utils';
-import { compileToAST, createASTCache } from '@8f4e/tokenizer';
-import { compileFunction } from './compileFunction';
-import { compileModules } from './compileModules';
-import { getError } from './compilerError';
-import { createInitialMemoryDataSegments } from './initialMemoryDataSegments';
-import {
-	assertUniqueModuleIds,
-	collectFunctionMetadataFromAsts,
-	collectNamespacesFromASTs,
-} from './semantic/buildNamespace';
-import { getCustomMemoryRegionName, validateMemoryRegionOptions } from './semantic/memoryRegions';
-import { expandMacros, parseMacroDefinitions } from './utils/macroExpansion';
-
-/** Expanded module source paired with cache and execution-entry metadata. */
-type ModuleCompilerSource = {
-	code: string[];
-	cacheKey: string;
-	entryName: string;
-};
+import { compileSubProgram, createCompilerCache } from './compileSubProgram';
+import { getCustomMemoryRegionName } from './semantic/memoryRegions';
 
 export { deriveEffectiveMemorySize } from '@8f4e/compiler-wasm-utils';
 export { serializeDiagnostic } from './diagnostic';
 
-/** Creates the default compiler cache used for validated AST reuse. */
-function createCompilerCache(): CompilerCache {
-	return {
-		ast: createASTCache<ValidatedAST>(),
-	};
-}
-
-/** Built-in WebAssembly export names that user functions are not allowed to reuse. */
-const RESERVED_EXPORT_NAMES = ['initDefaults'];
-
-/** Creates synthetic metadata for generated entry dispatcher functions. */
-function createEntryFunctionMetadata(
-	entryNames: readonly string[],
-	importedFunctionCount: number
-): FunctionMetadataLookup {
-	return Object.fromEntries(
-		entryNames.map((entryName, index) => [
-			entryName,
-			{
-				id: entryName,
-				signature: { parameters: [], returns: [] },
-				wasmIndex: importedFunctionCount + 1 + index,
-			} satisfies FunctionMetadata,
-		])
-	);
-}
-
-/** Calculates required byte size for each WebAssembly memory index. */
-function getRequiredMemoryBytesByIndex(
-	items: Array<{ memoryIndex: number; byteAddress?: number; wordAlignedSize?: number }>
-): Record<number, number> {
-	return items.reduce<Record<number, number>>((result, item) => {
-		const memoryIndex = item.memoryIndex;
-		const requiredBytes = (item.byteAddress ?? 0) + (item.wordAlignedSize ?? 0) * GLOBAL_ALIGNMENT_BOUNDARY;
-		result[memoryIndex] = Math.max(result[memoryIndex] ?? 0, requiredBytes);
-		return result;
-	}, {});
-}
-
-/** Converts non-default memory byte requirements into configured memory region names. */
-function getRequiredMemoryBytesByRegion(
-	requiredMemoryBytesByIndex: Record<number, number>,
-	memoryRegions: readonly string[]
-): Record<string, number> {
-	const result: Record<string, number> = {};
-
-	for (const [memoryIndexString, requiredBytes] of Object.entries(requiredMemoryBytesByIndex)) {
-		const memoryIndex = Number(memoryIndexString);
-		if (memoryIndex === 0 || requiredBytes <= 0) {
-			continue;
-		}
-
-		result[getCustomMemoryRegionName(memoryRegions, memoryIndex)] = requiredBytes;
-	}
-
-	return result;
-}
-
-/** Indexes prototype ASTs by id and rejects duplicate prototype declarations. */
-function collectPrototypeShapes(prototypes: readonly ValidatedPrototypeAST[]): Record<string, ValidatedPrototypeAST> {
-	const prototypeShapesById: Record<string, ValidatedPrototypeAST> = {};
-
-	for (const prototype of prototypes) {
-		const existing = prototypeShapesById[prototype.id];
-		if (existing) {
-			throw getError(ErrorCode.DUPLICATE_IDENTIFIER, prototype.prototypeLine, undefined, {
-				identifier: prototype.id,
-			});
-		}
-		prototypeShapesById[prototype.id] = prototype;
-	}
-
-	return prototypeShapesById;
-}
-
-/** Compiles modules, constants, functions, and prototypes into a complete WebAssembly program. */
+/** Compiles source input into a complete WebAssembly binary program. */
 export default function compile(
 	input: CompileInput,
 	options: CompileOptions,
 	cache = createCompilerCache()
 ): CompileResult {
-	validateMemoryRegionOptions(options);
-	const inputEntryNames = Object.keys(input.entries);
-	const entryModules = Object.entries(input.entries).flatMap(([entryName, modules]) =>
-		modules.map((module, index) => ({ entryName, module, index }))
-	);
-	const { constants, functions, prototypes, macros } = input;
-	const macroDefinitions = parseMacroDefinitions(macros);
-
-	const expandedPrototypes = prototypes.map((prototype, index) => {
-		return {
-			code: expandMacros(prototype, macroDefinitions),
-			cacheKey: `prototype:${index}`,
-		};
-	});
-
-	const astPrototypes = expandedPrototypes.map(
-		({ code, cacheKey }) => compileToAST(code, cache.ast, cacheKey) as ValidatedPrototypeAST
-	);
-	const prototypeShapesById = collectPrototypeShapes(astPrototypes);
-
-	// Expand macros in modules
-	const expandedModuleSources = entryModules.map(({ entryName, module, index }) => {
-		return {
-			code: expandMacros(module, macroDefinitions),
-			cacheKey: `entry:${entryName}:module:${index}`,
-			entryName,
-		};
-	}) satisfies ModuleCompilerSource[];
-
-	const expandedConstants = constants.map((constantsBlock, index) => {
-		return {
-			code: expandMacros(constantsBlock, macroDefinitions),
-			cacheKey: `constants:${index}`,
-		};
-	});
-
-	// Expand macros in functions
-	const expandedFunctions = functions.map((func, index) => {
-		return {
-			code: expandMacros(func, macroDefinitions),
-			cacheKey: `function:${index}`,
-		};
-	});
-
-	const astModuleEntries = expandedModuleSources.map(source => {
-		const ast = compileToAST(source.code, cache.ast, source.cacheKey) as ValidatedModuleAST;
-		return {
-			entryName: source.entryName,
-			ast,
-		};
-	});
-	const astConstants = expandedConstants.map(
-		({ code, cacheKey }) => compileToAST(code, cache.ast, cacheKey) as ValidatedConstantsAST
-	);
-	const entryNames = inputEntryNames;
-	const astModules = astModuleEntries.map(({ ast }) => ast);
-	const moduleEntryNames = astModuleEntries.map(({ entryName }) => entryName);
-	assertUniqueModuleIds(astModules);
-
-	const namespaceAsts = [...astModules, ...astConstants];
-	const namespaces = collectNamespacesFromASTs(
-		namespaceAsts,
-		GLOBAL_ALIGNMENT_BOUNDARY,
-		undefined,
-		astModules,
-		options,
-		prototypeShapesById
-	);
-
-	// Compile functions first with WASM indices and type registry
-	const astFunctions = expandedFunctions.map(
-		({ code, cacheKey }) => compileToAST(code, cache.ast, cacheKey) as ValidatedFunctionAST
-	);
-	const importedUserFunctionCount = astFunctions.filter(ast => ast.import).length;
-	const importedFunctionCount = importedUserFunctionCount;
-	const builtInFunctionCount = 1 + entryNames.length;
-	const userDefinedFunctionBaseIndex = importedFunctionCount + builtInFunctionCount;
-
-	// Collect pre-codegen function metadata so `call` target validation and
-	// function-body codegen can rely on the same registry before compilation finishes.
-	const entryFunctionMetadata = createEntryFunctionMetadata(entryNames, importedFunctionCount);
-	const userFunctionMetadata = collectFunctionMetadataFromAsts(astFunctions, {
-		importedFunctionBaseIndex: 0,
-		definedFunctionBaseIndex: userDefinedFunctionBaseIndex,
-		reservedFunctionIds: entryNames,
-		reservedExportNames: [...RESERVED_EXPORT_NAMES, ...entryNames],
-	});
-	const functionMetadata = { ...entryFunctionMetadata, ...userFunctionMetadata };
-
-	// Create a shared type registry for all functions
-	const functionTypeRegistry: FunctionTypeRegistry = {
-		types: [],
-		signatures: [],
-		baseTypeIndex: 3,
-	};
-
-	const compiledFunctions = astFunctions.map(ast =>
-		compileFunction(
-			ast,
-			namespaces,
-			functionMetadata[ast.id].wasmIndex,
-			functionTypeRegistry,
-			functionMetadata,
-			options
-		)
-	);
-	const importedUserFunctions = compiledFunctions.filter(func => func.import);
-	const definedFunctions = compiledFunctions.filter(func => !func.import);
-	const compiledFunctionsMap = Object.fromEntries(compiledFunctions.map(func => [func.id, func]));
-	const requiredMemoryBytesByIndexFromNamespaces = getRequiredMemoryBytesByIndex(Object.values(namespaces));
-	const totalModuleBytes = requiredMemoryBytesByIndexFromNamespaces[0] ?? 0;
-	const internalAllocator = { nextByteAddress: totalModuleBytes };
-
-	// Extract the unique function types and type indices from the registry
-	const uniqueUserFunctionTypes = functionTypeRegistry.types;
-	const userFunctionSignatureIndices = definedFunctions.map(func => func.typeIndex);
-
-	// Compile all modules in input order so editor layout can drive execution order.
-	const compiledModules = compileModules(
-		astModules,
-		{
-			...options,
-			startingMemoryWordAddress: 1,
-		},
-		namespaces,
-		functionMetadata,
-		internalAllocator,
+	const subProgram = compileSubProgram(input, options, cache);
+	const {
+		entryNames,
+		importedFunctionCount,
+		builtInFunctionCount,
+		compiledModules,
+		compiledModulesMap,
+		compiledFunctions,
+		compiledFunctionsMap,
+		importedUserFunctions,
+		definedFunctions,
 		functionTypeRegistry,
-		prototypeShapesById
-	).map((module, index) => ({
-		...module,
-		executionEntryName: moduleEntryNames[index],
-	}));
-
+		requiredMemoryBytesByIndex,
+		requiredMemoryBytes,
+		requiredMemoryBytesByRegion,
+		initialMemoryDataSegments,
+	} = subProgram;
 	const cycleFunctions = compiledModules.map(({ cycleFunction }) => cycleFunction);
 	const functionSignatures = compiledModules.map(() => 0x00);
-
-	// Calculate the required memory footprint from the compiled program.
-	const requiredMemoryBytesByIndexFromModules = getRequiredMemoryBytesByIndex(compiledModules);
-	const requiredMemoryBytes = Math.max(
-		requiredMemoryBytesByIndexFromModules[0] ?? 0,
-		internalAllocator.nextByteAddress
-	);
-	const requiredMemoryBytesByIndex: Record<number, number> = {
-		...requiredMemoryBytesByIndexFromModules,
-		0: requiredMemoryBytes,
-	};
-	const requiredMemoryBytesByRegion = getRequiredMemoryBytesByRegion(
-		requiredMemoryBytesByIndex,
-		options.memoryRegions ?? []
-	);
-	const maxUsedMemoryIndex = Math.max(0, ...Object.keys(requiredMemoryBytesByIndex).map(Number));
-
-	// Offset for user functions and module functions
+	const uniqueUserFunctionTypes = functionTypeRegistry.types;
+	const userFunctionSignatureIndices = definedFunctions.map(func => func.typeIndex);
 	const userFunctionCount = definedFunctions.length;
-	const getCompiledModuleFunctionIndex = (module: CompiledModule) =>
+	const getCompiledModuleFunctionIndex = (module: (typeof compiledModules)[number]) =>
 		importedFunctionCount + builtInFunctionCount + userFunctionCount + module.index;
 	const entryDispatcherFunctions = entryNames.map(entryName =>
 		createFunction(
@@ -308,8 +69,6 @@ export default function compile(
 			)
 		)
 	);
-
-	const initialMemoryDataSegments = createInitialMemoryDataSegments(compiledModules);
 	const memoryInitiatorFunction = [
 		...Object.entries(requiredMemoryBytesByIndex).flatMap(([memoryIndexString, requiredBytes]) => {
 			const memoryIndex = Number(memoryIndexString);
@@ -324,9 +83,7 @@ export default function compile(
 			...memoryInit(index, segment.memoryIndex),
 		]),
 	];
-
-	const compiledModulesMap = Object.fromEntries(compiledModules.map(module => [module.id, module]));
-
+	const maxUsedMemoryIndex = Math.max(0, ...Object.keys(requiredMemoryBytesByIndex).map(Number));
 	const memoryImports = Array.from({ length: maxUsedMemoryIndex + 1 }, (_, memoryIndex) => {
 		const requiredBytes = requiredMemoryBytesByIndex[memoryIndex] ?? 0;
 		const memorySizePages = Math.max(1, Math.ceil(requiredBytes / WASM_MEMORY_PAGE_SIZE));
@@ -345,7 +102,6 @@ export default function compile(
 			createFunctionImport(func.import!.moduleName, func.import!.fieldName, func.typeIndex)
 		),
 	];
-
 	const builtInFunctionSignatures = [0x00, ...entryNames.map(() => 0x00)];
 	const builtInFunctionBodies = [createFunction([], memoryInitiatorFunction), ...entryDispatcherFunctions];
 	const builtInExports = [
@@ -382,6 +138,6 @@ export default function compile(
 		compiledFunctions: compiledFunctionsMap,
 		requiredMemoryBytes,
 		...(Object.keys(requiredMemoryBytesByRegion).length > 0 ? { requiredMemoryBytesByRegion } : {}),
-		cache,
+		cache: subProgram.cache,
 	};
 }
